feat(server): move API server setup
