+++ bin/api-server/src/main.rs
@@ -2,12 +2,12 @@
 
 use std::net::SocketAddr;
 
-use api::run;
 use clap::Parser;
 use clickhouse::ClickhouseReader;
 use config::Opts;
 use dotenvy::dotenv;
 use primitives::shutdown::{ShutdownSignal, run_until_shutdown};
+use server::run;
 use tracing::info;
 use tracing_subscriber::filter::EnvFilter;
