+++ crates/server/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "server"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+api = { path = "../api" }
+clickhouse_lib = { path = "../clickhouse", package = "clickhouse" }
+axum.workspace = true
+tower-http.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+eyre.workspace = true
+
+[lints]
+workspace = true
